refactor(YALB-639): namespaced half of molecules
