+++ components/02-molecules/accordion/yds-accordion.twig
@@ -68,7 +68,7 @@
     {% endembed %}
     {% block accordion__items %}
       {% for item in accordion__items %}
-        {% include "@molecules/accordion/_accordion-item.twig" with {
+        {% include "@molecules/accordion/_yds-accordion-item.twig" with {
           accordion__item__heading: item.accordion__item__heading,
           accordion__item__content: item.accordion__item__content,
           accordion__item__heading__level: accordion__heading ? '3' : '2',
