+++ components/02-molecules/banner/banner.stories.js
@@ -1,6 +1,6 @@
 import tokens from '@yalesites-org/tokens/build/json/tokens.json';
 
-import bannerTwig from './cta/cta-banner.twig';
+import bannerTwig from './cta/yds-cta-banner.twig';
 
 import bannerData from './banner.yml';
