+++ components/02-molecules/callout/_yds-callout.twig
@@ -1,4 +1,4 @@
-{# Variables are documented in the main `callout.twig` file. #}
+{# Variables are documented in the main `yds-callout.twig` file. #}
 
 {% set callout__base_class = 'callout' %}
 {% set callout__link__type = callout__link__type|default('cta') %}
