+++ components/02-molecules/cards/reference-card/examples/_card--examples.twig
@@ -43,4 +43,4 @@
   {% endif %}
 {% endif %}
 
-{% include "@molecules/cards/reference-card/reference-card.twig" %}
+{% include "@molecules/cards/reference-card/yds-reference-card.twig" %}
