+++ components/02-molecules/image/_yds-content-image.scss
