+++ components/02-molecules/menu/_yds-menu-list.twig
@@ -3,7 +3,7 @@
 {% set menu__item %}
   {% block menu__item__content %}
     {% for item in items %}
-      {% include "@molecules/menu/_menu-item.twig" %}
+      {% include "@molecules/menu/_yds-menu-item.twig" %}
     {% endfor %}
   {% endblock %}
 {% endset %}
