+++ components/02-molecules/menu/yds-menu.twig
@@ -45,7 +45,7 @@
   {% if items %}
     {% set menu__modifiers = ['level-' ~ menu__level] %}
     {% block menu_list__content %}
-      {% include "@molecules/menu/_menu-list.twig" %}
+      {% include "@molecules/menu/_yds-menu-list.twig" %}
     {% endblock %}
   {% endif %}
 {% endmacro %}
