+++ components/02-molecules/molecules.scss
@@ -1,15 +1,15 @@
-@forward './accordion/accordion';
-@forward './alert/alert';
-@forward './banner/cta/cta-banner';
-@forward './callout/callout';
-@forward './cards/reference-card/reference-card';
-@forward './embed/embed';
-@forward './image/content-image';
-@forward './menu/menu';
-@forward './menu/menu-toggle/menu-toggle';
-@forward './meta/basic-meta/basic-meta';
-@forward './meta/event-meta/event-meta';
-@forward './page-title/page-title';
+@forward './accordion/yds-accordion';
+@forward './alert/yds-alert';
+@forward './banner/cta/yds-cta-banner';
+@forward './callout/yds-callout';
+@forward './cards/reference-card/yds-reference-card';
+@forward './embed/yds-embed';
+@forward './image/yds-content-image';
+@forward './menu/yds-menu';
+@forward './menu/menu-toggle/yds-menu-toggle';
+@forward './meta/basic-meta/yds-basic-meta';
+@forward './meta/event-meta/yds-event-meta';
+@forward './page-title/yds-page-title';
 @forward './pager/pager';
 @forward './pop-out-image/pop-out-image';
 @forward './pull-quote/pull-quote';
