+++ components/02-molecules/page-title/yds-page-title.twig
@@ -23,7 +23,7 @@
     } %}
     {% block page_title__meta %}
       {% if page_title__meta %}
-        {% include "@molecules/meta/basic-meta/basic-meta.twig" with {
+        {% include "@molecules/meta/basic-meta//yds-basic-meta.twig" with {
           basic_meta: page_title__meta,
         } %}
       {% endif %}
