+++ components/03-organisms/menu/primary-nav/primary-nav.twig
@@ -1,7 +1,7 @@
 {% set primary_nav__base_class = 'primary-nav' %}
 
 {# Menu #}
-{% include "@molecules/menu/menu.twig" with {
+{% include "@molecules/menu/yds-menu.twig" with {
   menu__base_class: primary_nav__base_class,
   menu__name: 'primary',
   items: primary_nav__items|default(items),
