+++ components/03-organisms/menu/utility-nav/utility-nav.twig
@@ -2,7 +2,7 @@
 
 <div {{ bem('utility-bar') }}>
   {# Menu #}
-  {% include "@molecules/menu/menu.twig" with {
+  {% include "@molecules/menu/yds-menu.twig" with {
     menu__base_class: utility_nav__base_class,
     menu__name: 'Utility',
     items: utility_nav__items|default(items),
