+++ components/03-organisms/site-header/site-header.twig
@@ -74,7 +74,7 @@
         link__blockname: site_header__base_class,
       } %}
       {# Mobile Menu Toggle #}
-      {% include "@molecules/menu/menu-toggle/menu-toggle.twig" %}
+      {% include "@molecules/menu/menu-toggle/yds-menu-toggle.twig" %}
     </div>
     <div {{ bem('menu-wrapper', [], site_header__base_class) }}>
       {# Secondary should come first on "desktop" sized screens #}
