+++ components/05-page-examples/events/event-grid.twig
@@ -10,7 +10,7 @@
     } %}
   {% endblock %}
   {% block page__content %}
-    {% include "@molecules/page-title/page-title.twig" %}
+    {% include "@molecules/page-title//yds-page-title.twig" %}
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '<p>This documentary examines the status of women during the transformation of Afghanistan from a totalitarian Taliban theocracy to a democratic country. It tells a story of Suraya Parlika, an Afghan woman who dared to enter public politics, which is eminently a territory of men. It also contains “black chronicle” of crimes from different provinces of Afghanistan, which was perpetrated by men against not only activists but also against women in their family, wives and daughters. Often just because they were determined to acquire a basic education.</p>',
     } %}
