+++ components/05-page-examples/miscellaneous/accordion-page.twig
@@ -1,15 +1,15 @@
 {% extends "@page-layouts/full-width.twig" %}
   {% block page__header %}
-    {% include "@molecules/alert/alert.twig" %}
+    {% include "@molecules/alert/yds-alert.twig" %}
     {% include "@organisms/site-header/site-header.twig" %}
     {% include "@organisms/menu/breadcrumbs/breadcrumbs.twig" %}
   {% endblock %}
   {% block page__content %}
-    {% include "@molecules/page-title/page-title.twig" %}
+    {% include "@molecules/page-title//yds-page-title.twig" %}
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '<p>Chemistry has been responsible for some of the most significant improvements in our quality of life over the last century.</p><p>The discovery of antibiotics and other pharmaceuticals, the advent of computers, and the development of industrial methods to produce fertilizer, all have required fundamental advances in chemistry. Chemistry is increasingly playing a central role in the development of alternative-energy vectors to replace fossil fuels, the realization of practical quantum computers, the discovery of new methods to treat and prevent diseases, and the adoption of more sustainable industrial processes.</p>',
     } %}
-    {% include "@molecules/accordion/accordion.twig" with {
+    {% include "@molecules/accordion/yds-accordion.twig" with {
       accordion__items: [
         {
           accordion__item__heading: accordion__item__heading,
@@ -32,7 +32,7 @@
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '<p>The undergraduate program in Chemistry at Yale reflects the position of chemistry as the foundation of all the molecular sciences. Students are equipped with the technical skills to appreciate the scientific basis for previous discoveries and develop the fundamental expertise required to make future breakthroughs. Under the tutelage of world-leading researchers, students are exposed to a broad range of topics in chemistry. The development of technical skills through lecture classes is complemented with hands-on experience in state-of-the-art chemistry laboratories.</p><p>Many students also perform independent laboratory research under the guidance of a faculty mentor. This rigorous training prepares students for professional careers in a diverse array of fields by teaching them how to apply the scientific method, providing them with skills in quantitative reasoning, and exposing them to scientific research.</p><p>After graduation, students with a B.A. or B.S. degree often pursue work or further studies in chemistry, biochemistry, or health-related disciplines, but also find their broad scientific training beneficial in energy research, policy, environment, business management, and law. As the problems of contemporary society involve ever more complex scientific issues, degree programs in the sciences become increasingly appropriate for students wishing to pursue careers in public policy, government, or public service.</p>',
     } %}
-    {% include "@molecules/accordion/accordion.twig" with {
+    {% include "@molecules/accordion/yds-accordion.twig" with {
       accordion__heading: 'Another Accordion Group',
       accordion__items: [
         {
