+++ components/05-page-examples/miscellaneous/qualtrics-embed.twig
@@ -1,12 +1,12 @@
 {% extends "@page-layouts/full-width.twig" %}
   {% block page__header %}
-    {% include "@molecules/alert/alert.twig" %}
+    {% include "@molecules/alert/yds-alert.twig" %}
     {% include "@organisms/site-header/site-header.twig" %}
     {% include "@organisms/menu/breadcrumbs/breadcrumbs.twig" %}
   {% endblock %}
   {% block page__content %}
-    {% include "@molecules/page-title/page-title.twig" %}
-    {% include "@molecules/embed/embed.twig" with {
+    {% include "@molecules/page-title//yds-page-title.twig" %}
+    {% include "@molecules/embed/yds-embed.twig" with {
       embed__src: 'https://yalesurvey.ca1.qualtrics.com/jfe/form/SV_cDezt2JVsNok77o',
       embed__title: 'Example Qualtics Form',
     } %}
