+++ components/05-page-examples/news/news-article.twig
@@ -10,14 +10,14 @@
     } %}
   {% endblock %}
   {% block page__content %}
-    {% include "@molecules/page-title/page-title.twig" %}
-    {% include "@molecules/image/content-image.twig" with {
+    {% include "@molecules/page-title//yds-page-title.twig" %}
+    {% include "@molecules/image/yds-content-image.twig" with {
       content_image__width: 'highlight',
     } %}
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '<p>Caitlin Davis, assistant professor of chemistry, has received a 2021 Scialog Collaborative Innovation Award to pursue research into the structure and function of enzyme filaments that regulate cell metabolism. Understanding cell processes is important as their alteration instigates the growth and spread of disease inside the body.</p><p>The award, from the Research Corporation for Science Advancement (RCSA) and the Gordon and Betty Moore Foundation, is part of RCSA’s Scialog Chemical Machinery of the Cell initiative to advance the fundamental understanding of chemical machinery and reactions in the intact cell. Teams of early career researchers who had not previously collaborated competed for seed funding for high-risk, high-reward projects to pursue novel lines of research.</p>',
     } %}
-    {% include "@molecules/image/content-image.twig" with {
+    {% include "@molecules/image/yds-content-image.twig" with {
       image__srcset: image__srcset__1,
       image__sizes: image__sizes__1,
       image__alt: image__alt__1,
